Render articles overview with new component accessible via articles route.
