+++ tests/scenarios/displaying-article-overview.test.js
@@ -2,7 +2,7 @@ var expect = require('chai').expect;
 var React = require('react/addons');
 var renderTarget, component;
 
-describe('displaying article overview', () => {
+xdescribe('displaying article overview', () => {
   var categories = [{id:1}, {id:2}];
   var articles = [{id:3, intensity: 3}, {id:4, intensity: 8}];
